pyproject.toml
[project]
name = "auto-blogger"
version = "0.1.0"
description = "AI-powered blog post generator using LangChain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
auto-blogger = "auto_blogger.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
